pyproject.toml
[project]
name = "urusai"
version = "0.0.1"
description = "agent that watches videos and refuses to answer without grounded evidence"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2",
    "pydantic-settings>=2",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "google-genai",
    "openai",
    "anthropic",
    "langgraph",
    "langchain-core",
    "ffmpeg-python",
    "scenedetect",
    "soundfile",
    "yt-dlp",
    "audio-separator",
    "pysrt",
    "webvtt-py",
    "networkx",
    "intervaltree",
    "faster-whisper",
]

[project.optional-dependencies]
ocr-local = [
    "paddleocr",
]
dev = [
    "pytest>=7",
    "pytest-asyncio",
    "ruff",
    "streamlit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/urusai"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
